+++ web/dashboard/src/app/api/v1/environments/[id]/components/[name]/ssr/route.ts
@@ -5,7 +5,7 @@ import { callSSR } from '@/lib/ssr/ssr';
 import crypto from 'crypto';
 import { LRUCache } from '@/lib/cache/LRUCache';
 
-const cache = new LRUCache<string, any>(100); // Set the cache capacity to 100
+const cache = new LRUCache<string, any>(10000); // Set the cache capacity to 100
 
 export const POST =
   async (
@@ -18,8 +18,9 @@ export const POST =
               status: HttpStatus.BAD_REQUEST,
           });
       }
-
+      console.time("authenticatedWithApiKeyUser");
       const user = await authenticatedWithApiKeyUser();
+      console.timeEnd("authenticatedWithApiKeyUser");
       const userId = user?.id;
       if (!userId) {
           return NextResponse.json('Api key not provided or invalid.', {
@@ -35,27 +36,35 @@ export const POST =
       }
 
       const req = await request.json();
+
+      console.time("get-env-component-config");
       const response = await getEnvironmentComponentConfig(id, name, userId);
+      console.timeEnd("get-env-component-config");
 
       const reqHash = crypto.createHash('sha256').update(JSON.stringify(req) + JSON.stringify(response)).digest('hex');
 
 
       if (cache.get(reqHash)) {
+          console.debug("Cache hit for hash:",name );
           return NextResponse.json(cache.get(reqHash), { status: HttpStatus.OK });
       }
+
+      console.debug("Cache miss for hash:", name );
 
       if (!response.success) {
           return NextResponse.json(response.error, {
               status: HttpStatus.BAD_REQUEST,
           });
       }
-      const {output, serverSideProps } = await callSSR(response.data.name, req, response.data.assets);
 
+      console.time("call-ssr");
+      const {output, serverSideProps } = await callSSR(response.data.name, req, response.data.assets);
+      console.timeEnd("call-ssr");
       if (output !== "") {
-          const result = {output, serverSideProps};
+          const result = {output, serverSideProps, version: response.data.version};
           cache.set(reqHash, result);
           return NextResponse.json(result, { status: HttpStatus.OK });
       }
 
-      return NextResponse.json({output, serverSideProps}, { status: HttpStatus.OK });
+      return NextResponse.json({output, serverSideProps, version: response.data.version}, { status: HttpStatus.OK });
   }