+++ web/dashboard/src/db/index.ts
@@ -1,5 +1,5 @@
-import { drizzle as drizzlePg, type PostgresJsDatabase } from 'drizzle-orm/postgres-js';
-import { drizzle as drizzleNeon, type NeonHttpDatabase } from 'drizzle-orm/neon-http';
+import { drizzle as drizzlePg } from 'drizzle-orm/postgres-js';
+import { drizzle as drizzleNeon } from 'drizzle-orm/neon-http';
 import postgres from 'postgres';
 import { neon } from '@neondatabase/serverless';
 import { remember } from '@epic-web/remember';
@@ -9,6 +9,9 @@ import * as projects from '@/data/projects/schema';
 import * as member from '@/data/member/schema';
 import * as components from '@/data/components/schema';
 import * as events from "@/data/events/schema";
+import { RedisCache } from "@/db/redis-cache";
+import {InMemoryCache} from "@/db/in-memory-cache";
+
 
 const schema = {
   ...users,
@@ -16,7 +19,11 @@ const schema = {
   ...member,
   ...components,
   ...events,
-}
+};
+
+const redisEnabled = process.env.DB_CACHE_STRATEGY === 'redis';
+
+const cache = remember("redis-cache", () => redisEnabled ? new RedisCache() : new InMemoryCache());
 
 function clientFactory() {
   let drizzle, client;
@@ -29,20 +36,21 @@ function clientFactory() {
     default:
       drizzle = drizzlePg;
       let connectionString =
-      client = postgres(
-        process.env.DRIZZLE_DATABASE_URL!,
-        {
-          max: process.env.DRIZZLE_DATABASE_MAX_CONNECTIONS ? parseInt(process.env.DRIZZLE_DATABASE_MAX_CONNECTIONS) : 25,
-          idle_timeout: 20,
-        }
-      );
+        client = postgres(
+          process.env.DRIZZLE_DATABASE_URL!,
+          {
+            max: process.env.DRIZZLE_DATABASE_MAX_CONNECTIONS ? parseInt(process.env.DRIZZLE_DATABASE_MAX_CONNECTIONS) : 25,
+            idle_timeout: 20,
+          },
+        );
   }
 
   return drizzle(client, {
+    cache,
     schema
   })
 }
 
 export const db = remember('db', () =>
-  clientFactory()
+  clientFactory(),
 );